requirements.txt
langchain_community 
langchain_core 
langchain_google_genai 
jq 
langchain 
langgraph
streamlit
python-dotenv